Put music search in separate component
