+++ client/src/components/MusicSelector/Detail.js
