+++ client/src/pages/Profile.js
@@ -1,17 +1,12 @@
-import React, { useState, useEffect }  from 'react';
-import Dropdown from '../components/Dropdown';
-import Listbox from '../components/Listbox';
-import Detail from '../components/Detail';
+import React  from 'react';
 
 import { Navigate, useParams } from 'react-router-dom';
 import { useQuery } from '@apollo/client';
 
 import { QUERY_SINGLE_PROFILE, QUERY_ME } from '../utils/queries';
 
+import MusicSearch from '../components/MusicSearch';
 import Auth from '../utils/auth';
-
-import { Credentials } from '../Credentials';
-import axios from 'axios';
 
 const Profile = () => {
   const { profileId } = useParams();
@@ -26,105 +21,9 @@ const Profile = () => {
   // Check if data is returning from the `QUERY_ME` query, then the `QUERY_SINGLE_PROFILE` query
   const profile = data?.me || data?.profile || {};
 
-  const spotify = Credentials();  
-
   console.log('RENDERING APP.JS');
 
-  // const data = [
-  //   {value: 1, name: 'A'},
-  //   {value: 2, name: 'B'},
-  //   {value: 3, name: 'C'},
-  // ]; 
-
-  const [token, setToken] = useState('');  
-  const [genres, setGenres] = useState({selectedGenre: '', listOfGenresFromAPI: []});
-  const [playlist, setPlaylist] = useState({selectedPlaylist: '', listOfPlaylistFromAPI: []});
-  const [tracks, setTracks] = useState({selectedTrack: '', listOfTracksFromAPI: []});
-  const [trackDetail, setTrackDetail] = useState(null);
-
-  useEffect(() => {
-
-    axios('https://accounts.spotify.com/api/token', {
-      headers: {
-        'Content-Type' : 'application/x-www-form-urlencoded',
-        'Authorization' : 'Basic ' + window.btoa(spotify.ClientId + ':' + spotify.ClientSecret)      
-      },
-      data: 'grant_type=client_credentials',
-      method: 'POST'
-    })
-    .then(tokenResponse => {      
-      setToken(tokenResponse.data.access_token);
-
-      axios('https://api.spotify.com/v1/browse/categories?locale=sv_US', {
-        method: 'GET',
-        headers: { 'Authorization' : 'Bearer ' + tokenResponse.data.access_token}
-      })
-      .then (genreResponse => {        
-        setGenres({
-          selectedGenre: genres.selectedGenre,
-          listOfGenresFromAPI: genreResponse.data.categories.items
-        })
-      });
-      
-    });
-
-  }, [genres.selectedGenre, spotify.ClientId, spotify.ClientSecret]); 
-
-  const genreChanged = val => {
-    setGenres({
-      selectedGenre: val, 
-      listOfGenresFromAPI: genres.listOfGenresFromAPI
-    });
-
-    axios(`https://api.spotify.com/v1/browse/categories/${val}/playlists?limit=10`, {
-      method: 'GET',
-      headers: { 'Authorization' : 'Bearer ' + token}
-    })
-    .then(playlistResponse => {
-      setPlaylist({
-        selectedPlaylist: playlist.selectedPlaylist,
-        listOfPlaylistFromAPI: playlistResponse.data.playlists.items
-      })
-    });
-
-    console.log(val);
-  }
-
-  const playlistChanged = val => {
-    console.log(val);
-    setPlaylist({
-      selectedPlaylist: val,
-      listOfPlaylistFromAPI: playlist.listOfPlaylistFromAPI
-    });
-  }
-
-  const buttonClicked = e => {
-    e.preventDefault();
-
-    axios(`https://api.spotify.com/v1/playlists/${playlist.selectedPlaylist}/tracks?limit=10`, {
-      method: 'GET',
-      headers: {
-        'Authorization' : 'Bearer ' + token
-      }
-    })
-    .then(tracksResponse => {
-      setTracks({
-        selectedTrack: tracks.selectedTrack,
-        listOfTracksFromAPI: tracksResponse.data.items
-      })
-    });
-  }
-
-  const listboxClicked = val => {
-
-    const currentTracks = [...tracks.listOfTracksFromAPI];
-
-    const trackInfo = currentTracks.filter(t => t.track.id === val);
-
-    setTrackDetail(trackInfo[0].track);
-  }
-
-
+  // Use React Router's `<Navigate />` component to redirect to personal profile page if username is yours
   if (Auth.loggedIn() && Auth.getProfile().data._id === profileId) {
     return <Navigate to="/me" />;
   }
@@ -142,27 +41,12 @@ const Profile = () => {
     );
   }
 
-
   return (
     <div>
       <h2 className="card-header">
         {profileId ? `${profile.name}'s` : 'Your'} PLAYLIST is here...
       </h2>
-
-      <form onSubmit={buttonClicked}>        
-          <Dropdown label="Genre :" options={genres.listOfGenresFromAPI} selectedValue={genres.selectedGenre} changed={genreChanged} />
-          <Dropdown label="Playlist :" options={playlist.listOfPlaylistFromAPI} selectedValue={playlist.selectedPlaylist} changed={playlistChanged} />
-          <div className="col-sm-6 row form-group px-0">
-            <button type='submit' className="btn btn-success col-sm-12">
-              Search
-            </button>
-          </div>
-          <div className="row">
-            <Listbox items={tracks.listOfTracksFromAPI} clicked={listboxClicked} />
-            {trackDetail && <Detail {...trackDetail} /> }
-          </div>        
-      </form>
-
+      <MusicSearch/>
     </div>
   );
 };
